+++ src/main/java/Controller.java
@@ -1,4 +1,2 @@
-package sample;
-
 public class Controller {
 }
