+++ src/main/java/Module.java
@@ -1,6 +1,3 @@
-package sample;
-
-
 public class Module extends Subject {
