+++ src/main/java/ModuleInterface.java
@@ -1,5 +1,3 @@
-package sample;
-
 import javafx.application.Platform;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
@@ -53,71 +51,74 @@ public class ModuleInterface extends Parent {
 
 
     public ModuleInterface() {
-        Platform.runLater(() -> {
+        Platform.runLater(new Runnable() {
+            @Override
+            public void run() {
 
 
-            System.out.println("Constructor of ModuleInterface");
+                System.out.println("Constructor of ModuleInterface");
 
 
-            //------LAYOUTS-------
+                //------LAYOUTS-------
 
 
-            //Layout Page
-            VBox vBox = new VBox(10);
+                //Layout Page
+                VBox vBox = new VBox(10);
 
 
-            //Layout Title line
-            HBox hBoxTitleButton = new HBox(30);
-            hBoxTitleButton.setAlignment(Pos.CENTER_LEFT);
+                //Layout Title line
+                HBox hBoxTitleButton = new HBox(30);
+                hBoxTitleButton.setAlignment(Pos.CENTER_LEFT);
 
-            //Layout Subtitles
-            HBox hBoxSubtitle = new HBox(30);
-            hBoxSubtitle.setAlignment(Pos.BOTTOM_LEFT);
+                //Layout Subtitles
+                HBox hBoxSubtitle = new HBox(30);
+                hBoxSubtitle.setAlignment(Pos.BOTTOM_LEFT);
 
-            //Layout Subjects
-            hBoxMarks.setPadding(new Insets(10, 10, 10, 10));
-
-
-            //-------TEXTFIELD--------
-
-            TextField moduleName = new TextField();
-            moduleName.setStyle("-fx-font: 30 berlin; -fx-font-weight: bold;");
-            moduleName.setText(module.getMarkName());
-
-            moduleName.textProperty().addListener((observable, oldValue, newValue) -> module.setMarkName(moduleName.getText()));
-
-            moduleName.setStyle("-fx-font: 20 berlin; -fx-font-weight: bold;");
-            moduleName.setPrefWidth(220);
+                //Layout Subjects
+                hBoxMarks.setPadding(new Insets(10, 10, 10, 10));
 
 
-            //------AVERAGE LINE------
+                //-------TEXTFIELD--------
 
-            MarkInterface moduleAverage = new MarkInterface(module);
+                TextField moduleName = new TextField();
+                moduleName.setStyle("-fx-font: 30 berlin; -fx-font-weight: bold;");
+                moduleName.setText(module.getMarkName());
 
+                moduleName.textProperty().addListener((observable, oldValue, newValue) -> module.setMarkName(moduleName.getText()));
 
-            //------BUTTONS------
-
-            Button newSubjectButton = new Button("Add Subject");
-
-            newSubjectButton.setOnAction(e -> {
-
-                subjectInterfaceArrayList.add(new SubjectInterface(module));
-                generateHBox(subjectInterfaceArrayList);
-
-            });
+                moduleName.setStyle("-fx-font: 20 berlin; -fx-font-weight: bold;");
+                moduleName.setPrefWidth(220);
 
 
-            //-------ADD DEFAULT EXAM------
+                //------AVERAGE LINE------
 
-            //subjectInterfaceArrayList.add(new SubjectInterface(module));
-            //generateHBox(subjectInterfaceArrayList);
+                MarkInterface moduleAverage = new MarkInterface(module);
 
 
-            //--------ADD ELEMENTS TO LAYOUTS------
+                //------BUTTONS------
 
-            hBoxTitleButton.getChildren().addAll(moduleName, newSubjectButton);
-            vBox.getChildren().addAll(hBoxTitleButton, moduleAverage, hBoxMarks, hBoxSubtitle);
-            this.getChildren().add(vBox);
+                Button newSubjectButton = new Button("Add Subject");
+
+                newSubjectButton.setOnAction(e -> {
+
+                    subjectInterfaceArrayList.add(new SubjectInterface(module));
+                    ModuleInterface.this.generateHBox(subjectInterfaceArrayList);
+
+                });
+
+
+                //-------ADD DEFAULT EXAM------
+
+                //subjectInterfaceArrayList.add(new SubjectInterface(module));
+                //generateHBox(subjectInterfaceArrayList);
+
+
+                //--------ADD ELEMENTS TO LAYOUTS------
+
+                hBoxTitleButton.getChildren().addAll(moduleName, newSubjectButton);
+                vBox.getChildren().addAll(hBoxTitleButton, moduleAverage, hBoxMarks, hBoxSubtitle);
+                ModuleInterface.this.getChildren().add(vBox);
+            }
         });
 
     }
