+++ src/main/java/Read.java
@@ -1,5 +1,3 @@
-package sample;
-
 import java.io.File;
 import java.io.FileReader;
