+++ src/main/java/Save.java
@@ -1,5 +1,3 @@
-package sample;
-
 import java.io.PrintWriter;
 import java.util.ArrayList;
