+++ src/main/java/Subject.java
@@ -1,5 +1,3 @@
-package sample;
-
 import java.util.ArrayList;
 
 public class Subject extends Mark {
